#97: Rename ServiceBrokerMetrics to ServiceBrokerMetricsService
