+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/metrics/BindingMetricsServiceSpec.groovy
@@ -27,7 +27,7 @@ class BindingMetricsServiceSpec extends Specification {
     private ServiceInstanceRepository serviceInstanceRepository
     private LastOperationRepository lastOperationRepository
     private ServiceBindingRepository serviceBindingRepository
-    private BindingMetricsService bindingMetricsService
+    private BindingMetricsServiceService bindingMetricsService
     private CFServiceRepository cfServiceRepository
     private PlanRepository planRepository
     private MeterRegistry meterRegistry
@@ -45,7 +45,7 @@ class BindingMetricsServiceSpec extends Specification {
         cfServiceRepository.findAll() >> new ArrayList<CFService>()
         serviceBindingRepository.findAll() >> new ArrayList<ServiceBinding>()
 
-        bindingMetricsService = new BindingMetricsService(serviceInstanceRepository, cfServiceRepository, lastOperationRepository, serviceBindingRepository, planRepository, meterRegistry)
+        bindingMetricsService = new BindingMetricsServiceService(serviceInstanceRepository, cfServiceRepository, lastOperationRepository, serviceBindingRepository, planRepository, meterRegistry)
     }
 
     def "retrieve total nr of successful service bindings"() {
