+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/metrics/ProvisionRequestsMetricsServiceSpec.groovy
@@ -32,7 +32,7 @@ class ProvisionRequestsMetricsServiceSpec extends Specification {
     private LastOperationRepository lastOperationRepository
     private PlanRepository planRepository
     private MeterRegistry meterRegistry
-    private ProvisionRequestsMetricsService provisionRequestsMetricsService
+    private ProvisionRequestsMetricsServiceService provisionRequestsMetricsService
 
     def setup() {
         serviceInstanceRepository = Mock(ServiceInstanceRepository)
@@ -44,7 +44,7 @@ class ProvisionRequestsMetricsServiceSpec extends Specification {
         cfServiceRepository.findAll() >> new ArrayList<CFService>()
         planRepository.findAll() >> new ArrayList<Plan>()
 
-        provisionRequestsMetricsService = new ProvisionRequestsMetricsService(serviceInstanceRepository, cfServiceRepository, lastOperationRepository, planRepository, meterRegistry)
+        provisionRequestsMetricsService = new ProvisionRequestsMetricsServiceService(serviceInstanceRepository, cfServiceRepository, lastOperationRepository, planRepository, meterRegistry)
     }
 
     def "retrieve total nr of provision Requests"() {
