+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/metrics/ProvisionedInstancesMetricsServiceSpec.groovy
@@ -32,7 +32,7 @@ class ProvisionedInstancesMetricsServiceSpec extends Specification {
     private LastOperationRepository lastOperationRepository
     private PlanRepository planRepository
     private MeterRegistry meterRegistry
-    private ProvisionedInstancesMetricsService provisioningMetricsService
+    private ProvisionedInstancesMetricsServiceService provisioningMetricsService
 
     def setup() {
         serviceInstanceRepository = Mock(ServiceInstanceRepository)
@@ -44,7 +44,7 @@ class ProvisionedInstancesMetricsServiceSpec extends Specification {
         cfServiceRepository.findAll() >> new ArrayList<CFService>()
         planRepository.findAll() >> new ArrayList<Plan>()
 
-        provisioningMetricsService = new ProvisionedInstancesMetricsService(serviceInstanceRepository, cfServiceRepository, lastOperationRepository, planRepository, meterRegistry)
+        provisioningMetricsService = new ProvisionedInstancesMetricsServiceService(serviceInstanceRepository, cfServiceRepository, lastOperationRepository, planRepository, meterRegistry)
     }
 
     def "retrieve total nr of provisioned instances"() {
